Add bundle version and exported packages.
Note that the package versions are at 1.0 as they should follow the OSGi
semantic versioning scheme. More information about that can be found
here: www.osgi.org/wiki/uploads/Links/SemanticVersioning.pdf
In short: if the interfaces don't change the packages stay at 1.0. Only
if the interfaces in a particular package will the package version be
affected.
Also updated the dependency on the maven bundle plugin to version 2.3.5
(latest).