Fix Javadoc Warnings
This commit fixes all the Javadoc warnings I have when building
locally. These all {@link } references that can not be resolved and
result in no hyperlink being generated in the resulting HTML.

The cause seem to be two different limitations in the Javadoc tool
 * `package-info.html` apparently always needs fully qualified class
    names even if they are in the same package or in `java.lang`.
 * Referencig nested types (`Cache.Entry`) in Javadoc does not work
   when only the outer type (`Cache.Entry`) is imported. You could
   import the nested type but when it is only referenced qualified in
   the code then the IDE is likely to generated an unused import
   warning.

I did sign the Terracotta Contributor Agreement.