Add License Information to Artifacts
There's currently no license information in the POM or JAR. It is my
understanding that this usually prevents syncing to Maven Central [1].

This pull request:

 * adds a license section to the POM
 * adds a license file to the final artifact, taken from [2]

The license name is derived from what servlet API uses ("SUN LIMITED
EVALUATION LICENSE").

I did not add a `Bundle-License` section in the `bnd.bnd` file since
the Maven Bundle Plugin should generate this automatically from the
POM.

I did not add a `<distribution>repo</distribution>` because
I didn't see a section in the license that would allow me to
redistribute the software. In my limited understanding this makes the
whole syncing with Maven Central questionable. If that is indeed the
case then `<distribution>manual</distribution>` should be added.

Once the JSR goes final the license has to be updated with the final
license template.

Obviously somebody from Terracotta legal should have a look a this.

I did sign the Terracotta Contributor Argreement.

 [1] http://blog.sonatype.com/people/2010/01/nexus-oss-ecosystem/
 [2] http://jcp.org/ja/resources/license_reference